requirements.txt

Flask==2.3.3
python-telegram-bot==20.7
